requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.11b1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
fake-useragent==0.1.11
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
lxml==4.7.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
oauthlib==3.1.1
outcome==1.1.0
parse==1.19.0
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.19.1
pycparser==2.21
pyee==8.2.2
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyppeteer==0.2.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt6==6.2.2
PyQt6-Qt6==6.2.2
PyQt6-sip==13.2.0
pyquery==1.4.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
tomli==1.2.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.4.0
typing_extensions==4.0.1
urllib3==1.26.7
w3lib==1.22.0
webdriver-manager==3.5.2
websockets==9.1
wsproto==1.0.0
zipp==3.6.0
